--- sass/components/_label-common.scss
@@ -1,23 +1,0 @@
-.Label {
-    position: relative;
-    font-family: "Open Sans", sans-serif;
-    font-size: 14/10*1rem;
-    line-height: (18/14);
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    overflow: hidden;
-
-    &.Label___small {
-        font-size: 12/10*1rem;
-        line-height: (16/12);
-    }
-    
-    &.Label___large {
-        font-size: 16/10*1rem;
-        line-height: (20/16);
-    }
-
-    &.Label___wrap {
-        white-space: normal;
-    }
-}
